Idmanager and refractor code. (#113)
* Added IdManager. To be used for storage of Last Unique Id used by Task,Customer and Driver.

* Added JsonAdaptedDriver

* Added JsonAdaptedDriver.

* Refractor AddCommand and EditCommand for Driver and Customer to take in unique id.

* Updated DeleteIdCommand, setDriver method in Task.
-setDriver in Task will now adjust Status and EventTime accordingly.
-DeleteIdCommand will now prevent deletion of driver or customer who are already allocated to a task.

* Updated EditTaskCommand.
-if edited date is before today, then don't allow.

* Updated after review.

+++ src/main/java/seedu/address/logic/commands/DeleteIdCommand.java
@@ -31,8 +31,9 @@ public class DeleteIdCommand extends Command {
             + PREFIX_CUSTOMER + " 2";
 
     public static final String MESSAGE_INVALID_ID = "Invalid %1$s id.";
-
     public static final String MESSAGE_DELETE_PERSON_SUCCESS = "Deleted %1$s: %2$s";
+    public static final String MESSAGE_CANNOT_DELETE_IF_ALLOCATED =
+            "This %1$s has already been allocated to a Task, therefore, cannot delete";
 
     private final String className;
     private final int id;
@@ -72,15 +73,27 @@ public class DeleteIdCommand extends Command {
             }
 
             Customer customerToDelete = model.getCustomer(id);
+
+            //if customer is already allocated to a task, whether complete or incomplete
+            if (model.hasTaskBelongsToCustomer(customerToDelete)) {
+                throw new CommandException(String.format(MESSAGE_CANNOT_DELETE_IF_ALLOCATED, className));
+            }
+
             model.deleteCustomer(customerToDelete);
             return new CommandResult(String.format(MESSAGE_DELETE_PERSON_SUCCESS, className, customerToDelete));
         } else {
             //deletion for Driver
             if (!model.hasDriver(id)) {
-                throw new CommandException(String.format(MESSAGE_DELETE_PERSON_SUCCESS, className));
+                throw new CommandException(String.format(MESSAGE_CANNOT_DELETE_IF_ALLOCATED, className));
             }
 
             Driver driverToDelete = model.getDriver(id);
+
+            //if driver is already allocated to a task, whether complete or incomplete
+            if (model.hasTaskBelongsToDriver(driverToDelete)) {
+                throw new CommandException(String.format(MESSAGE_DELETE_PERSON_SUCCESS, className));
+            }
+
             model.deleteDriver(driverToDelete);
             return new CommandResult(String.format(MESSAGE_DELETE_PERSON_SUCCESS, className, driverToDelete));
         }
